# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Kanho API
  description: Kanban project management API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh access token
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /cards/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      operationId: getCard
      summary: Get card detail
      tags: [cards]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Card detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDetail"
    patch:
      operationId: updateCard
      summary: Update card fields
      tags: [cards]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCardRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
    delete:
      operationId: deleteCard
      summary: Delete card permanently
      tags: [cards]
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Deleted
  /cards/{id}/move:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    post:
      operationId: moveCard
      summary: Move card to column + position
      tags: [cards]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveCardRequest"
      responses:
        "200":
          description: Moved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /cards/{id}/archive:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    post:
      operationId: archiveCard
      summary: Archive card
      tags: [cards]
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Archived
...
  /attachments/presign:
    post:
      operationId: presignUpload
      summary: Get presigned upload URL
      tags: [attachments]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresignRequest"
      responses:
        "200":
          description: Presigned URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresignResponse"
  /search/cards:
    get:
      operationId: searchCards
      summary: Full-text search cards
      tags: [cards]
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: workspace_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    Priority:
      type: string
      enum: [none, low, medium, high, urgent]
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    AuthResponse:
      type: object
      required: [access_token, refresh_token, user]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        avatar_url:
          type: string
        created_at:
          type: string
          format: date-time
...
    Card:
      type: object
      required: [id, project_id, board_id, column_id, card_number, title, priority, position, created_by, created_at]
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        board_id:
          type: string
          format: uuid
        column_id:
          type: string
          format: uuid
        card_number:
          type: integer
        readable_id:
          type: string
          description: "e.g. PROJ-42"
        title:
          type: string
        priority:
          $ref: "#/components/schemas/Priority"
        position:
          type: string
        start_date:
          type: string
          format: date-time
        due_date:
          type: string
          format: date-time
        archived_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        assignees:
          type: array
          items:
            $ref: "#/components/schemas/User"
        labels:
          type: array
          items:
            $ref: "#/components/schemas/Label"
    CardDetail:
      allOf:
        - $ref: "#/components/schemas/Card"
        - type: object
          properties:
            description:
              type: string
            description_json:
              type: object
            comments:
              type: array
              items:
                $ref: "#/components/schemas/Comment"
            attachments:
              type: array
              items:
                $ref: "#/components/schemas/Attachment"
...
    UpdateCardRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        description_json:
          type: object
        priority:
          $ref: "#/components/schemas/Priority"
        start_date:
          type: string
          format: date-time
        due_date:
          type: string
          format: date-time
        assignee_ids:
          type: array
          items:
            type: string
            format: uuid
        label_ids:
          type: array
          items:
            type: string
            format: uuid
    MoveCardRequest:
      type: object
      required: [column_id, position]
      properties:
        column_id:
          type: string
          format: uuid
        position:
          type: string
    Comment:
      type: object
      required: [id, card_id, author_id, content, created_at]
      properties:
        id:
          type: string
          format: uuid
        card_id:
          type: string
          format: uuid
        author_id:
          type: string
          format: uuid
        author:
          $ref: "#/components/schemas/User"
        content:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Attachment:
      type: object
      required: [id, card_id, filename, content_type, size_bytes, storage_key, uploaded_by, created_at]
      properties:
        id:
          type: string
          format: uuid
        card_id:
          type: string
          format: uuid
        filename:
          type: string
        content_type:
          type: string
        size_bytes:
          type: integer
          format: int64
        storage_key:
          type: string
        url:
          type: string
        uploaded_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
...
    PresignRequest:
      type: object
      required: [filename, content_type]
      properties:
        filename:
          type: string
        content_type:
          type: string
    PresignResponse:
      type: object
      required: [presigned_url, storage_key, public_url]
      properties:
        presigned_url:
          type: string
        storage_key:
          type: string
        public_url:
          type: string
    Label:
      type: object
      required: [id, workspace_id, name, color, created_at]
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: string
        created_at:
          type: string
          format: date-time
...